pyproject.toml
[project]
name = "uws-core"
version = "1.0.0"
description = "Universal Web Scraper core"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
  "playwright>=1.47,<1.56",
  "pydantic==2.9.2",
  "pyyaml==6.0.2",
  "pandas==2.2.2",
  "python-dotenv==1.0.1",
  "tqdm==4.66.5"
]

[project.scripts]
uws-run = "main:main"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
